+++ ckanext/base/fanstatic/js/color_picker.js
