+++ ckanext/base/fanstatic/resource.config
