+++ ckanext/opendata_theme/controller.py
@@ -17,53 +17,10 @@ from ckan.plugins.toolkit import (
 )
 
 from ckanext.opendata_theme.constants import LAYOUTS
-from ckanext.opendata_theme.processors import custom_style_processor, custom_naming_processor
-from ckanext.opendata_theme.processor import custom_style_processor as csp
+from ckanext.opendata_theme.processors import custom_naming_processor
 
 
 class CustomCSSController(admin.AdminController):
-    def custom_css(self):
-        extra_vars = {}
-        if request.method == 'POST':
-            data = clean_dict(dict_fns.unflatten(
-                tuplize_dict(parse_params(request.POST))))
-
-            custom_css, css_metadata = csp.get_custom_css(data)
-            contrast_errors = csp.check_contrast()
-            extra_vars.update({'errors': contrast_errors})
-            if not contrast_errors:
-                self.save_css_metadata(custom_css, css_metadata)
-                css_metadata = self.sort_inputs_by_position(css_metadata)
-                extra_vars.update(self.split_inputs_onto_two_columns(css_metadata))
-                redirect_to(
-                    controller='ckanext.opendata_theme.controller:CustomCSSController',
-                    action='custom_css',
-                    extra_vars=extra_vars
-                )
-
-        css_metadata = self.get_custom_css_metadata()
-
-        if not css_metadata:
-            _, css_metadata = csp.get_custom_css({})
-            self.save_css_metadata({}, css_metadata)
-        else:
-            css_metadata = ast.literal_eval(css_metadata)
-        css_metadata = self.sort_inputs_by_position(css_metadata)
-        extra_vars.update(self.split_inputs_onto_two_columns(css_metadata))
-        return render('admin/custom_css.html', extra_vars=extra_vars)
-
-    def reset_custom_css(self):
-        extra_vars = {}
-        _, css_metadata = csp.get_custom_css({})
-        self.save_css_metadata({}, css_metadata)
-        css_metadata = self.sort_inputs_by_position(css_metadata)
-        extra_vars.update(self.split_inputs_onto_two_columns(css_metadata))
-        redirect_to(
-            controller='ckanext.opendata_theme.controller:CustomCSSController',
-            action='custom_css',
-            extra_vars=extra_vars
-        )
-
     def custom_home_page(self):
         extra_vars = {}
         if request.method == 'POST':
@@ -92,7 +49,7 @@ class CustomCSSController(admin.AdminController):
             custom_naming = custom_naming_processor.get_custom_naming({})
             get_action('config_option_update')({}, {"ckanext.opendata_theme.custom_naming": custom_naming})
         else:
-           custom_naming = ast.literal_eval(custom_naming)
+            custom_naming = ast.literal_eval(custom_naming)
         custom_naming = self.sort_inputs_by_position(custom_naming)
 
         # Get last or default layout
@@ -119,22 +76,6 @@ class CustomCSSController(admin.AdminController):
             action='custom_home_page',
             extra_vars=extra_vars
         )
-
-    @staticmethod
-    def save_css_metadata(custom_css, css_metadata):
-        get_action('config_option_update')({}, {"ckanext.opendata_theme.custom_raw_css": custom_css})
-        get_action('config_option_update')({}, {"ckanext.opendata_theme.custom_css_metadata": css_metadata})
-
-    @staticmethod
-    def get_custom_css_metadata():
-        return get_action('config_option_show')({}, {"key": "ckanext.opendata_theme.custom_css_metadata"})
-
-    @staticmethod
-    def split_inputs_onto_two_columns(data):
-        input_numbers = len(data)
-        part_1 = OrderedDict(list(data.items())[0:input_numbers / 2])
-        part_2 = OrderedDict(list(data.items())[input_numbers / 2:])
-        return {"data_part_1": part_1, "data_part_2": part_2}
 
     @staticmethod
     def sort_inputs_by_position(css_metadata):
